backend/requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.29.0
bcrypt==5.0.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.7
dnspython==2.7.0
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet==3.0.3
h11==0.16.0
idna==3.11
motor==3.7.1
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
PyJWT==2.8.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.27
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
